requirements.txt
Django==4.0
djangorestframework==3.15.1
PyMySQL==1.1.0
django-simple-history==3.5.0